requirements.txt
numpy==2.4.4 ; python_version >= "3.13"
pandas==3.0.2 ; python_version >= "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
tzdata==2026.1 ; python_version >= "3.13" and (sys_platform == "win32" or sys_platform == "emscripten")
httpx==0.27.0
selectolax==0.3.21
pydantic==2.7.1
tenacity==8.2.3
pandas==2.2.2
pyarrow==15.0.2
loguru==0.7.2
psycopg2-binary==2.9.9